requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
# google-cloud-speech==2.25.1
google-cloud-aiplatform==1.48.0
python-multipart==0.0.9
firebase-admin

